# packages/contracts/openapi.v1.yaml
openapi: 3.1.0
info:
  title: The Penthouse API
  version: 1.0.0-alpha
  description: Private social platform API. All protected endpoints require a Bearer access token.
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          oneOf:
            - type: string
              description: Human-readable error message
            - type: object
              description: Zod field-level validation error (fieldErrors + formErrors)
...
    ChatSummary:
      type: object
      required: [id, type, name, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [dm, channel]
        name:
          type: string
        updatedAt:
          type: string
          format: date-time
    Message:
      type: object
      required: [id, chatId, senderId, content, type, createdAt]
      properties:
        id:
          type: string
          format: uuid
        chatId:
          type: string
          format: uuid
        senderId:
          type: string
          format: uuid
        senderUsername:
          type: string
          description: Immutable login identifier. May be omitted in future member-facing clients.
        senderDisplayName:
          type: string
        senderAvatarUrl:
          type: string
          nullable: true
        content:
          type: string
        type:
          type: string
          enum: [text, image, video, gif, file]
        metadata:
          type: object
          nullable: true
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        clientMessageId:
          type: string
          description: Client-supplied idempotency key (optional on received messages)
    SendMessageRequest:
      type: object
      required: [content, type, clientMessageId]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 4000
        type:
          type: string
          enum: [text, image, video, gif, file]
        metadata:
          type: object
          nullable: true
          additionalProperties: true
        clientMessageId:
          type: string
          minLength: 8
          maxLength: 128
          description: Idempotency key supplied by client. Same key = same message (deduped).
    SendMessageResponse:
      type: object
      required: [message, deduped]
      properties:
        message:
          $ref: '#/components/schemas/Message'
        deduped:
          type: boolean
          description: True if clientMessageId matched an existing message (idempotent replay)
    UploadResponse:
      type: object
      required: [id, fileName, originalFileName, url, size, mediaKind]
      properties:
        id:
          type: string
          format: uuid
        fileName:
          type: string
        originalFileName:
          type: string
        url:
          type: string
        size:
          type: integer
          description: File size in bytes
        contentType:
          type: string
        mediaKind:
          type: string
          enum: [image, video, file]
    GifResult:
      type: object
      required: [id, url, previewUrl, provider]
      properties:
        id:
          type: string
        url:
          type: string
        previewUrl:
          type: string
        title:
          type: string
          nullable: true
        width:
          type: integer
          nullable: true
        height:
          type: integer
          nullable: true
        provider:
          type: string
          enum: [giphy, klipy]
    GifSearchResponse:
      type: object
      required: [provider, results]
      properties:
        provider:
          type: string
          enum: [giphy, klipy]
        results:
          type: array
          items:
            $ref: '#/components/schemas/GifResult'
...
paths:
...
  /api/v1/chats:
    get:
      summary: List chats available to current user
      operationId: listChats
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Array of chat summaries ordered by last activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatSummary'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: 'Unauthorized'
  /api/v1/chats/{chatId}/messages:
    get:
      summary: List messages for chat (cursor-paginated, newest first)
      operationId: listMessages
      security:
        - bearerAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: cursor
          in: query
          required: false
          description: Message ID to paginate before (exclusive)
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          description: Number of messages to return (1–100, default 30)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 30
      responses:
        '200':
          description: Array of messages in descending chronological order
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not a member of this chat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: 'Forbidden'

    post:
      summary: Send message with idempotent clientMessageId
      operationId: sendMessage
      security:
        - bearerAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '200':
          description: Message sent or deduped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendMessageResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not a member of this chat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: 'Forbidden'
  /api/v1/media/upload:
    post:
      summary: Upload media (multipart/form-data)
      operationId: uploadMedia
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: File to upload (max configurable via UPLOAD_MAX_MB env)
      responses:
        '201':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          description: No file provided or file too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/gifs/{provider}/trending:
    get:
      summary: Get trending GIFs from a configured provider
      operationId: getTrendingGifs
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
            enum: [giphy, klipy]
      responses:
        '200':
          description: Trending GIF results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GifSearchResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: GIF provider not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/gifs/{provider}/search:
    get:
      summary: Search GIFs from a configured provider
      operationId: searchGifs
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
            enum: [giphy, klipy]
        - in: query
          name: q
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Matching GIF results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GifSearchResponse'
        '400':
          description: Missing search query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: GIF provider not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'